Fix code-review findings in the electronic-mode and reverse-rate work
Memory safety:

* Ambipolar electrons live in the scratch elist array, not in
  particle->particles, so they have no per-particle custom storage.  The
  guards on that storage tested p->ispecies == ambispecies, but the
  reacting E-disposal runs after the reaction has already overwritten the
  electron's species with the heavy product species, so the guard stopped
  firing and zero_elec()/relax_electronic_mode()/the discrete-vibmode
  sampler indexed the custom arrays with a difference between two
  unrelated allocations.  Replace the species test with a has_custom()
  pointer test in both CollideVSS and CollideVSSKokkos.

* An electron promoted out of elist into the master particle list reuses
  a slot whose custom data belongs to whatever particle held it last;
  add_particle() does not clear it (unlike the add_particle(...)
  overload).  Zero the slot in both ambipolar collision paths and in the
  Kokkos twin.

Hangs and undefined behavior:

* Particle::erot() and ParticleKokkos::erot() divided by rottemp[0],
  which is 0 until a rotfile is read.  Collide::init() rejects that
  combination, but create_particles runs first, so tratio went infinite,
  jmax was an undefined int conversion, and the rejection loop never
  terminated.  Guard the divisor instead.

Kokkos/CPU parity:

* CollideVSSKokkos::init() re-implements the Collide::init() checks and
  had dropped three of them: the discrete-rotation rotational-temperature
  check, the "no species has electronic data" check, and the ban on
  electronic states for the ambipolar electron species.  Without the last
  one the Kokkos path wrote per-particle electronic data for every elist
  electron on every collision.

* FixElecmodeKokkos's Kokkos copy constructor never set copy = 1, leaving
  the flag indeterminate (Fix(SPARTA*) has an empty body), so the
  destructor could run remove_custom() on an already-deleted Particle.

* ParticleKokkos::update_elec_views() zero-filled only three of its six
  mirrors; the views are allocated with Kokkos::NoInit and
  create_mirror_view() does not initialize, so d_elecstates,
  d_elec_default_rels and d_enforce_spin_conservation held host garbage
  outside each species' populated range.

Correctness:

* Particle::bisectTelec() took the particle count as int while its only
  call site passes a double tally (the Kokkos mirror already used
  double), truncating fractional counts and overflowing under
  fix ave/grid ave running.

* Particle::electronic_distribution_func() skipped its whole body when
  collide->elecstyle was not DISCRETE and returned the buffer untouched,
  which callers dereference unconditionally.  The distribution is a pure
  function of the species electronic data, so always compute it and error
  if the species has none.

* Particle::add_species() omitted elecfile from the keyword loop that
  sizes names[], so it was counted as a species ID.

+++ src/particle.h
@@ -202,7 +202,7 @@ class Particle : protected Pointers {
   int ielec(int, double, class RanKnuth *);
   double* electronic_distribution_func(int, double);
   double elec_energy(int isp, double temp_elec);
-  double bisectTelec(int isp, double eelec, int count);
+  double bisectTelec(int isp, double eelec, double count);
 
   void write_restart_species(FILE *fp);
   void read_restart_species(FILE *fp);
